requirements.txt
torch==2.5.1
torchvision==0.20.1
easyocr
supervision==0.18.0
transformers
ultralytics==8.3.50
numpy
opencv-python
opencv-python-headless
accelerate
timm
einops==0.8.0
azure-identity
dill
paddlepaddle
paddleocr
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-community==0.3.5
python-dotenv
pyautogui
keyboard
watchdog
moviepy
pillow
minecraft_launcher_lib
fastapi
uvicorn